+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -1,5 +1,6 @@
 package dinkplugin;
 
+import dinkplugin.domain.AccountType;
 import dinkplugin.domain.AchievementDiary;
 import dinkplugin.domain.ChatNotificationType;
 import dinkplugin.domain.ChatPrivacyMode;
@@ -189,12 +190,14 @@ public interface DinkPluginConfig extends Config {
     )
     String externalSection = "External Plugin Requests";
 
+    /*
     @ConfigSection(
         name = "Leagues",
         description = "Settings for notifying when you complete league tasks, unlock areas, and redeem relics",
         position = 200,
         closedByDefault = true
     )
+     */
     String leaguesSection = "Leagues";
 
     @ConfigSection(
@@ -499,6 +502,18 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "deniedAccountTypes",
+        name = "Denied Account Types",
+        description = "Types of accounts that should not trigger notifications.<br/>" +
+            "Has no effect if 'RSN Filter Mode' is set to 'Exclusively Allow'",
+        position = 1021,
+        section = advancedSection
+    )
+    default Set<AccountType> deniedAccountTypes() {
+        return EnumSet.noneOf(AccountType.class);
+    }
+
+    @ConfigItem(
         keyName = "discordWebhook", // do not rename; would break old configs
         name = "Primary Webhook URLs",
         description = "The default webhook URL to send notifications to, if no override is specified.<br/>" +
@@ -724,7 +739,8 @@ public interface DinkPluginConfig extends Config {
         description = "If non-empty, Leagues messages are sent to this URL, instead of the primary URL.<br/>" +
             "Note: this only applies to the Leagues notifier, not every notifier in a seasonal world",
         position = 0,
-        section = webhookSection
+        section = webhookSection,
+        hidden = true
     )
     default String leaguesWebhook() {
         return "";
@@ -2081,7 +2097,8 @@ public interface DinkPluginConfig extends Config {
         name = "Enable Leagues",
         description = "Enable notifications upon various leagues events",
         position = 200,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean notifyLeagues() {
         return false;
@@ -2092,7 +2109,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Image",
         description = "Send image with the notification",
         position = 201,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesSendImage() {
         return true;
@@ -2103,7 +2121,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Area Unlocks",
         description = "Send notifications upon area unlocks",
         position = 202,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesAreaUnlock() {
         return true;
@@ -2114,7 +2133,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Relic Unlocks",
         description = "Send notifications upon relic unlocks",
         position = 203,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesRelicUnlock() {
         return true;
@@ -2125,7 +2145,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Completed Tasks",
         description = "Send notifications upon completing a task",
         position = 204,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesTaskCompletion() {
         return true;
@@ -2136,7 +2157,8 @@ public interface DinkPluginConfig extends Config {
         name = "Send Mastery Unlocks",
         description = "Send notifications upon combat mastery selections",
         position = 205,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default boolean leaguesMasteryUnlock() {
         return true;
@@ -2147,7 +2169,8 @@ public interface DinkPluginConfig extends Config {
         name = "Task Min Difficulty",
         description = "The minimum tier of a task for a notification to be sent",
         position = 206,
-        section = leaguesSection
+        section = leaguesSection,
+        hidden = true
     )
     default LeagueTaskDifficulty leaguesTaskMinTier() {
         return LeagueTaskDifficulty.EASY;