+++ src/main/java/dinkplugin/notifiers/BaseNotifier.java
@@ -2,10 +2,13 @@ package dinkplugin.notifiers;
 
 import dinkplugin.DinkPluginConfig;
 import dinkplugin.SettingsManager;
+import dinkplugin.domain.FilterMode;
 import dinkplugin.domain.SeasonalPolicy;
 import dinkplugin.message.DiscordMessageHandler;
 import dinkplugin.message.NotificationBody;
+import dinkplugin.util.Utils;
 import dinkplugin.util.WorldUtils;
+import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
 import net.runelite.api.Player;
 import net.runelite.api.WorldType;
@@ -14,6 +17,7 @@ import org.apache.commons.lang3.StringUtils;
 import javax.inject.Inject;
 import java.util.Set;
 
+@Slf4j
 public abstract class BaseNotifier {
 
     @Inject
@@ -36,8 +40,22 @@ public abstract class BaseNotifier {
         if (WorldUtils.isIgnoredWorld(world)) {
             return false;
         }
+        return accountPassesConfig();
+    }
+
+    /**
+     * @return whether the player name passes the configured filtered RSNs and the account type is permitted
+     */
+    protected boolean accountPassesConfig() {
         Player player = client.getLocalPlayer();
-        return player != null && settingsManager.isNamePermitted(player.getName());
+        if (player == null) {
+            return false;
+        }
+        if (config.nameFilterMode() != FilterMode.ALLOW && config.deniedAccountTypes().contains(Utils.getAccountType(client))) {
+            log.info("Skipping notification from {} due to denied account type", getClass().getSimpleName());
+            return false;
+        }
+        return settingsManager.isNamePermitted(player.getName());
     }
 
     protected abstract String getWebhookUrl();