+++ src/main/java/dinkplugin/notifiers/PetNotifier.java
@@ -386,7 +386,7 @@ public class PetNotifier extends BaseNotifier {
             entry("Baby chinchompa", new SkillSource(Skill.HUNTER, 82_758, 315)), // black chinchompas
             entry("Baby mole", new KcSource("Giant Mole", 1.0 / 3_000)),
             entry("Baron", new KcSource("Duke Sucellus", 1.0 / 2_500)),
-            entry("Bran", new KcSource("Royal Titans", null)), // unknown drop rate
+            entry("Bran", new MultiKcSource("Branda the Fire Queen", 1.0 / 3_000, "Eldric the Ice King", 1.0 / 3_000)), // assume loot not sacrificed
             entry("Butch", new KcSource("Vardorvis", 1.0 / 3_000)),
             entry("Beaver", new SkillSource(Skill.WOODCUTTING, 264_336, 85)), // teaks
             entry("Bloodhound", new KcSource("Clue Scroll (master)", 1.0 / 1_000)),
@@ -490,7 +490,7 @@ public class PetNotifier extends BaseNotifier {
             entry("Pet dark core", new KcSource("Corporeal Beast", 1.0 / 5_000)),
             entry("Pet general graardor", new KcSource("General Graardor", 1.0 / 5_000)),
             entry("Pet k'ril tsutsaroth", new KcSource("K'ril Tsutsaroth", 1.0 / 5_000)),
-            entry("Pet kraken", new KcSource("Kraken", 1.0 / 5_000)),
+            entry("Pet kraken", new KcSource("Kraken", 1.0 / 3_000)),
             entry("Pet penance queen", new Source() {
                 @Override
                 Double getProbability(Client client, KillCountService kcService) {