+++ src/main/java/dinkplugin/util/KillCountService.java
@@ -53,7 +53,7 @@ public class KillCountService {
 
     public static final Set<Integer> SPECIAL_LOOT_NPC_IDS = Set.of(
         NpcID.THE_WHISPERER, NpcID.THE_WHISPERER_12205, NpcID.THE_WHISPERER_12206, NpcID.THE_WHISPERER_12207,
-        NpcID.ARAXXOR, NpcID.ARAXXOR_13669, 14148, 14149
+        NpcID.ARAXXOR, NpcID.ARAXXOR_13669, NpcID.BRANDA_THE_FIRE_QUEEN_14148, NpcID.ELDRIC_THE_ICE_KING_14149
     );
     public static final Set<String> SPECIAL_LOOT_NPC_NAMES = Set.of("The Whisperer", "Araxxor", "Branda the Fire Queen", "Eldric the Ice King");
 
@@ -161,8 +161,8 @@ public class KillCountService {
             String cacheKey = getCacheKey(LootRecordType.UNKNOWN, boss);
             killCounts.asMap().merge(cacheKey, kc - 1, Math::max);
 
-            if (boss.equals(GAUNTLET_BOSS) || boss.equals(CG_BOSS) || boss.startsWith(TOA) || boss.startsWith(TOB) || boss.startsWith(COX)) {
-                // populate lastDrop to workaround loot tracker quirks
+            if (boss.equals("Araxxor") || boss.equals(GAUNTLET_BOSS) || boss.equals(CG_BOSS) || boss.startsWith(TOA) || boss.startsWith(TOB) || boss.startsWith(COX)) {
+                // populate lastDrop to workaround loot tracking quirks
                 this.lastDrop = new Drop(boss, LootRecordType.EVENT, Collections.emptyList());
 
                 if (!ConfigUtil.isPluginDisabled(configManager, RL_LOOT_PLUGIN_NAME)) {
@@ -318,8 +318,11 @@ public class KillCountService {
             case PLAYER:
                 return "player_" + sourceName;
             default:
+                // exceptions where boss name in chat message differs from npc name
+                if ("Whisperer".equals(sourceName)) return "The Whisperer";
                 if ("The Gauntlet".equals(sourceName)) return GAUNTLET_BOSS;
                 if (CG_NAME.equals(sourceName)) return CG_BOSS;
+
                 return sourceName;
         }
     }